feat(main.go): add new file to load content from S3 using Pulumi
This new file is added to load content from S3 using Pulumi. It includes
the main function and several helper functions to create a Kubernetes
job that fetches data from S3. The job uses environment variables for
the access key and secret key, which are fetched from Kubernetes
secrets. The job also uses a list of S3 buckets specified in the Pulumi
configuration. This allows for a flexible and secure way to load data
from S3 into the application.