chore(Pulumi.staging.yaml): update namespace and secret values from 'dev' to 'staging'
The namespace and secret values were updated to reflect the change in
the environment from 'dev' to 'staging'. This is necessary to ensure
that the correct resources are accessed in the staging environment. The
tag value was also updated to 'staging-4806391' to indicate the new
version of the application for the staging environment.