feat(container.go): add new container specification file for Kubernetes deployment
This new file, container.go, is added to define the container
specifications for the Kubernetes deployment. It includes functions to
create the container specification, environment variables, arguments,
and port specifications. This is crucial for the deployment of the
application in a Kubernetes environment, as it allows for the
configuration of the container's environment, arguments, and exposed
ports.