feat(Pulumi.dev.yaml): add new configuration file for Pulumi in development environment
This new configuration file is added to manage the infrastructure of the
backend application in a development environment using Pulumi. It
includes the secrets provider, encrypted key, and configurations for
different backend applications such as annotation, content, order, and
stock. This allows for better management and configuration of the
development environment.