feat: add .kodiak.toml configuration file for automating PR merges
The .kodiak.toml file has been added to automate the process of merging
pull requests. This will help in automatically merging PRs opened by
"dependabot" that are "minor" or "patch" version upgrades. "major"
version upgrades will be ignored. This will help in keeping the
dependencies up-to-date and reduce the manual effort required to merge
these PRs.