feat(service.go): add new service specification functions for Kubernetes deployment
The new functions in service.go allow for the creation of service
specifications for Kubernetes deployments. This includes the creation of
service arguments, service specification arguments, and service port
specifications. These functions will be used to automate and streamline
the deployment process, making it more efficient and less prone to human
error.