feat(deployment.go): add Kubernetes deployment specification for frontend application
The new file, deployment.go, contains the Kubernetes deployment
specification for the frontend application. This includes the
deploymentSpec function that returns the deployment arguments, the
deploymentSpecArgs function that returns the deployment specification
arguments, and the deploymentLabelSelector function that returns the
label selector arguments. The deploymentPodTemplate function is also
added to return the pod template specification arguments. This addition
allows for the deployment of the frontend application in a Kubernetes
environment, providing scalability and manageability.