feat(Pulumi.yaml): add new Pulumi configuration file for frontend application
A new Pulumi configuration file has been added to manage the
infrastructure of the frontend application. This file will allow us to
use Pulumi's Infrastructure as Code (IaC) capabilities to automate the
provisioning and management of our frontend application's
infrastructure. The runtime specified is Go, indicating that the
infrastructure code will be written in Go.