feat(volume.go): add CreateContainerVolumeMount function to create Kubernetes volume mounts
The new function CreateContainerVolumeMount is added to the k8s package.
This function is used to create Kubernetes volume mounts. It takes a
volume and a mountPath as arguments and returns a VolumeMountArgs
object. This function will help in creating volume mounts for Kubernetes
containers, which is a common requirement in Kubernetes deployments.