chore(Pulumi.dev.yaml, Pulumi.staging.yaml): update image tags and log level
The image tags for the publication and stock-center services in both the
development and staging environments have been updated to reflect the
latest changes in the codebase. This ensures that the deployed services
are up-to-date with the latest features and bug fixes. Additionally, the
log level in the staging environment has been changed from 'info' to
'debug' to provide more detailed logging information, which can be
useful for troubleshooting issues in the staging environment.