refactor(container.go): move VolumeMounts from createRepoContainerSpec to baseContainerSpec
The VolumeMounts property was moved from the createRepoContainerSpec
function to the baseContainerSpec function. This change was made to
improve the code structure and make it more readable. Now, all the base
arguments for the container specification are defined in one place,
which makes the code easier to understand and maintain.