feat(volume.go): add CreateVolumeSpec and volumeItems functions to create volume specifications
The CreateVolumeSpec function was added to create volume specifications
with secretName, volumeName, and items as parameters. This function will
be useful in creating volume specifications for Kubernetes deployments.
The volumeItems function was also added to convert VolumeItemsProperties
into KeyToPathArray. This function will be used by CreateVolumeSpec to
convert the items parameter into the required format. These changes will
make it easier to create and manage volume specifications in Kubernetes
deployments.