refactor(container.go): rename dumpCmd to dumpCmdTmpl and refactor pulumi.ToStringArray to pulumi.StringArray
The variable name dumpCmd was changed to dumpCmdTmpl to better reflect
its purpose as a template for the dump command. The pulumi.ToStringArray
function was replaced with pulumi.StringArray to simplify the code and
improve readability. This change does not affect the functionality of
the code but makes it easier to understand and maintain.