feat(job.go): add support for PostgreSQL backup jobs
This change introduces two new functions,
createPostgresPodTemplateSpecArgs and createPostgresCronJobSpec. These
functions allow the creation of a Kubernetes Pod and CronJob
specifically for backing up PostgreSQL databases. This is a necessary
feature for users who want to automate their PostgreSQL database
backups. The Pod created by createPostgresPodTemplateSpecArgs includes a
container running a PostgreSQL backup process and a volume for storing
the backup data. The CronJob created by createPostgresCronJobSpec runs
this Pod on a schedule defined by the user.