feat(main.go): add function to create GCP storage bucket
The function 'createGcpBucket' has been added to create a Google Cloud
Platform (GCP) storage bucket. This function is necessary to automate
the process of creating a storage bucket in GCP, which is a crucial part
of the backup application. The bucket is created in the 'US-CENTRAL1'
location with versioning enabled, providing a way to keep multiple
variants of an object in the same bucket.