feat(just_modules/image.justfile): create new justfile for managing machine images in GCP
The addition of the `gcp-image` module in the Justfile and the creation
of a corresponding justfile in `just_modules/image.justfile` allows for
automated management of machine images in Google Cloud Platform. This
setup provides a structured and reusable way to create, manage, and
monitor machine images, enhancing the deployment processes and backup
strategies for instances, specifically tailored for Kubernetes nodes.