feat(cloudnative-pg-operator): add Pulumi configuration files for deploying PostgreSQL operator
Introduce Pulumi.dev.yaml and Pulumi.yaml to set up and configure the
Cloud Native PostgreSQL operator using Helm charts in a Kubernetes
environment. This setup specifies the use of a specific version of the
PostgreSQL operator and its deployment in the 'dev' namespace, ensuring
that the development environment is ready for database operations.