feat(arangodb): add secret management for ArangoDB deployment configuration
Introduce a new Secret struct in ArangoDeploymentConfig to manage
sensitive information such as passwords. Implement the creation of
a Kubernetes Secret during the installation of the ArangoDB deployment.
This change enhances security by allowing sensitive data to be stored
in a Kubernetes Secret rather than hardcoded in the application.
Additionally, the ArangoDB deployment now references this Secret for
bootstrap configuration, improving the overall security posture of
the deployment.