feat(service.go): add CreateService method to GraphqlServer struct for creating a Kubernetes service
The CreateService method is added to the GraphqlServer struct in the service.go file. This method is responsible for creating a Kubernetes service based on the provided configuration. It generates the service name, deployment name, and sets up the necessary metadata and specifications for the service. This addition enhances the functionality of the GraphqlServer by enabling it to create and manage Kubernetes services.