feat(container.go): add functions to generate environment variable arrays for ConfigMap and Secret
The functions ConfigMapEnvArgsArray, SecretEnvArgsArray, and ContainerEnvArgsArray have been added to the EventMessengerEmail struct in container.go. These functions generate arrays of environment variables for ConfigMap and Secret, and combine them into a single array respectively. This change enhances the modularity and reusability of the code by separating the logic for generating different types of environment variables.