refactor(deployment.go): modify CreateDeployment and CreateDeploymentSpec functions to accept claimName parameter
The CreateDeployment function in deployment.go now accepts a claimName parameter in addition to the pulumi.Context parameter. This change allows for more flexibility and customization when creating a deployment. The CreateDeploymentSpec function has also been updated to include a volume with a persistent volume claim based on the provided claimName. This enhancement improves the modularity and extensibility of the deployment creation process.